Extract ML Job status helpers to separate file
