+++ x-pack/legacy/plugins/siem/public/components/ml_popover/jobs_table/job_switch.tsx
@@ -8,6 +8,7 @@ import styled from 'styled-components';
 import React, { useState, useCallback } from 'react';
 import { EuiFlexGroup, EuiFlexItem, EuiLoadingSpinner, EuiSwitch } from '@elastic/eui';
 import { SiemJob } from '../types';
+import { isJobLoading, isJobStarted, isJobFailed } from '../../ml/helpers';
 
 const StaticSwitch = styled(EuiSwitch)`
   .euiSwitch__thumb,
@@ -23,23 +24,6 @@ export interface JobSwitchProps {
   isSiemJobsLoading: boolean;
   onJobStateChange: (job: SiemJob, latestTimestampMs: number, enable: boolean) => Promise<void>;
 }
-
-// Based on ML Job/Datafeed States from x-pack/legacy/plugins/ml/common/constants/states.js
-const enabledStates = ['started', 'opened'];
-const loadingStates = ['starting', 'stopping', 'opening', 'closing'];
-const failureStates = ['deleted', 'failed'];
-
-export const isChecked = (jobState: string, datafeedState: string): boolean => {
-  return enabledStates.includes(jobState) && enabledStates.includes(datafeedState);
-};
-
-export const isJobLoading = (jobState: string, datafeedState: string): boolean => {
-  return loadingStates.includes(jobState) || loadingStates.includes(datafeedState);
-};
-
-export const isFailure = (jobState: string, datafeedState: string): boolean => {
-  return failureStates.includes(jobState) || failureStates.includes(datafeedState);
-};
 
 export const JobSwitchComponent = ({
   job,
@@ -64,8 +48,8 @@ export const JobSwitchComponent = ({
         ) : (
           <StaticSwitch
             data-test-subj="job-switch"
-            disabled={isFailure(job.jobState, job.datafeedState)}
-            checked={isChecked(job.jobState, job.datafeedState)}
+            disabled={isJobFailed(job.jobState, job.datafeedState)}
+            checked={isJobStarted(job.jobState, job.datafeedState)}
             onChange={handleChange}
             showLabel={false}
             label=""
